requirements.txt
streamlit
langchain
google-generativeai
crewai
pint
python-dotenv